+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+<a name="2.3.0"></a>
+
+# [2.3.0](https://github.com/deckgo/deckdeckgo/compare/v2.1.0...v2.2.0) (In progress...)
+
+### Applications
+
+- docs: v1.9.1 ([CHANGELOG](https://github.com/deckgo/deckdeckgo/blob/master/docs/CHANGELOG.md))
+- remote: v1.6.2 ([CHANGELOG](https://github.com/deckgo/deckdeckgo/blob/master/remote/CHANGELOG.md))
+
+### Web Components
+
+- highlight-code: v1.1.4 ([CHANGELOG](https://github.com/deckgo/deckdeckgo/blob/master/webcomponents/highlight-code/CHANGELOG.md))
+
+### Others
+
+- gatsby-remark-highlight-code: v1.4.7 ([CHANGELOG](https://github.com/deckgo/gatsby-remark-highlight-code/blob/master/CHANGELOG.md))
+- starter kit: v2.6.5 ([CHANGELOG](https://github.com/deckgo/deckdeckgo-starter/blob/master/CHANGELOG.md))
+
 <a name="2.2.0"></a>
 
 # [2.2.0](https://github.com/deckgo/deckdeckgo/compare/v2.1.0...v2.2.0) (2020-08-17)